# Shortcut in ParetoCheck add/check for performance [changelog skip]