@slvlirnoff 2022-04-06
Also I've seen that the Heuristic when pruning after relaxing transit is usually faster (less `leftDominanceExist` check per `qualify`) than the check done after relaxing transfers. This is due to the transfer not taking into account the time to wait for the next departure in the heuristic, having then a much lower potential `endTime`. Nothing we can do about this I think.

(I might try later to go through the list backward instead, to see if this improves overall performance.)