@slvlirnoff 2022-04-06
Iterating backward in the `qualify` method yield an additional small improvement.

It's probably possible for the `add` call too, but needs slightly more complex logic in order to then remove the other dominated results and might not be as important. Let me know if you want me to try to do this too.

```
Test case ids       : [  1   2   3   4  5   6   7   8   9  10  11  12 13  14  15  16 17  18  19  20  21  22   23  24  25   26  27  28  29  30  31   32  33   34   35   36  101  102 103 104  105  106  107]
Number of paths     : [  2   2   3   2  1   1   3   1   3   3   3   3  2   2   3   2  3   2   1   3   2   2    3   1   1    3   2   3   1   3   3    3   3    3    3    3    3    3   3   3    3    3    2]
Transit times(ms)   : [278 261 224 294 90 167 407 312 140 417 195 104 84 274 163 105 87 375 284 105 166 378 1374 368 883 1325 370 403 160 600 678 1069 794 2928 1089 2234 1280  926 746 638  790 1087 1606]
Total times(ms)     : [280 261 272 312 90 167 540 312 150 417 195 112 92 277 173 105 88 376 284 107 166 378 1375 369 884 1325 370 411 160 601 679 1075 797 2929 1090 2235 1699 1042 786 734 1261 1503 3128]
Successful searches : 41 / 43
Sample              : 4 / 4
Time total          : 30 seconds

Worker: 
 ==> mc_destination : [  527,  498,  492,  496 ] Avg: 503.3

Total:  
 ==> mc_destination : [  733,  688,  683,  689 ] Avg: 698.3
 ```