@t2gran 2022-04-07
> Also I've seen that the Heuristic when pruning after relaxing transit is usually faster (less leftDominanceExist check per qualify) than the check done after relaxing transfers. This is due to the transfer not taking into account the time to wait for the next departure in the heuristic, having then a much lower potential endTime. Nothing we can do about this I think.

The heuristics is not at all very tight at the moment, this is on our Road-map to fix - we need it to add another criteria to McRaptor. We want to be able to support a competition neutral search and add a criteria for "emulated-price". To get a much faster heuristics we can pre-calculate dwell and hop times. The current heuristic search should be replaced by one "single-criteria" search for each criteria (arrival-time, num-of-transfers, generalized-cost).