@slvlirnoff 2022-04-07
Also regarding the heuristic, on the first raptor minute iteration, the 'paths' are empty. For a long search where the minimum journey would be 4-5 transfers, it means that the whole network will be explored for 4-5 rounds with no destination arrival path to prune against.

Similarly for searches where the solution has a low frequency or high amount of waiting time, the heuristic will perform less and less well as we iterate through the raptor minutes until we find a new valid path that can prune the exploration.

What could be potentially done is to pick the best itineraries from the heuristic search and use these to 'prepopulate' the DestinationArrivalPaths.

That's how it behave now (using my poor ascii drawing skill 😄 )
```
1st iteration:
paths: <empty>
raptor:
                              o------> o------> o-----> o-------> o-----> best itinerary found at iteration 1, round 5
                              (completely unrelated journey)
                              o------> o----> o-----> o----------> o----->( + heuristic min_time --->) valid exploration path
```

```
after 5 rounds of the 1st iteration and on later iterations
paths: 
 1)                             o------> o------> o-----> o-------> o----->
raptor:
                              (completely unrelated journey)
                               o------> (+ heuristic min_time o----> o-----> o--->)  x rejected early, because dominated by 1)
```


```
a few iterations later:
paths: 
 1)                             o------> o------> o-----> o-------> o----->
raptor:
              (rather unrelated journey, potentially valid compared to 1)
              o------> o----> o----->( + heuristic min_time --->) valid exploration path compared to 1)
              
a few iterations later:
        o------> o------> o-----> o-------> o-----> next best itinerary found 
```

If instead we had:

```
Prepopulated paths: 
 1)                             o------> o------> o-----> o-------> o----->
 2)        o------> o------> o-----> o-------> o----->
 3) o------> o------> o-----> o-------> o----->
raptor:
              (all previous unrelated journey)
              o------> (+ heuristic min_time o----> o-----> o--->)  x rejected early, because dominated
              
      o--------> unique very fast journey, not found by heuristic, kept
      
paths: 
 1)                             o------> o------> o-----> o-------> o----->
 2)        o------> o------> o-----> o-------> o----->
 3)   o--------> (replace previously prepopulated path)
```

(BTW: these schema illustrate why changing the 'qualify' check to a reverse order make sense, as the last path are more likely to dominate the exploration path)

To expand the itineraries found in the heuristics, I can think of different approaches:
-  the heuristic best paths could be inserted in the destination arrival paths
- whenever a path is inserted in the destination arrival paths, it could be expanded over the search window by iterating over the stop patterns times
- we could run an initial mc raptor search to find potential solutions that uses only the stop pattern identified by the heuristic

I've logged the time spent per raptor iteration, and it's interesting to see the 'spikes' at specific iterations.
But as you can see the first raptor minute iteration takes twice as much time as any of the heuristics, I guess this can be explained by the above. Later iterations takes between 0ms to 100-200ms.
```
Fwd heuristic:
1st Iteration done: 28800 in 128ms

Backward heuristic:
1st Iteration done: 175475 in 187ms

RangeRaptor
1st Iteration done: 32340 in 441ms
```